@@ -3,7 +3,7 @@ package com.accountbook.domain;
 import com.accountbook.request.PostEdit;
 import lombok.*;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Getter
