@@ -1,11 +1,18 @@
 package com.accountbook.service;
 
 import com.accountbook.domain.Siginup;
+import com.accountbook.domain.User;
+import com.accountbook.exception.AlreadyExistsEmailException;
 import com.accountbook.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
+
+import javax.swing.text.html.Option;
+import java.rmi.AlreadyBoundException;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -13,9 +20,23 @@ public class AuthService {
 
     private final UserRepository userRepository;
 
+    private final PasswordEncoder passwordEncoder;
 
-    @Transactional
+
     public void signup(Siginup signup){
+        Optional<User> userOptional = userRepository.findByEmail(signup.getEmail());
+        if (userOptional.isPresent()){
+            throw new AlreadyExistsEmailException();
+        }
 
+        String encryptedPassword = passwordEncoder.encode(signup.getPassword());
+
+        var user = User.builder()
+                .email(signup.getEmail())
+                .name(signup.getName())
+                .password(encryptedPassword)
+                        .build();
+
+        userRepository.save(user);
     }
 }
