@@ -5,6 +5,10 @@ import com.accountbook.domain.Siginup;
 import com.accountbook.service.AuthService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.password.NoOpPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
@@ -19,9 +23,13 @@ public class AuthController {
 
     private final APPConfig appConfig;
 
+    @GetMapping("/auth/login")
+    public String login(){
+        return "login page";
+    }
+
     @PostMapping("/auth/signup")
     public void signup(@RequestBody Siginup signup){
         authService.signup(signup);
     }
-
 }
